reverse_engineer_API/requirements.txt
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.1
seaborn==0.12.2
Office365-REST-Python-Client==2.5.1
msal==1.25.0
O365==2.0.33
geopandas==0.13.2
flask==3.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1
pandas>=1.3.0 